log (feature): Scala Native support
